+++ tests/helpers.py
@@ -4,8 +4,8 @@ from argparse import Namespace
 from typing import Iterable
 
 import torch
-from transformers import BertConfig, BertForSequenceClassification
 import transformers
+from transformers import BertConfig, BertForSequenceClassification
 from transformers.optimization import AdamW
 from transformers.tokenization_utils import PreTrainedTokenizerBase
 
@@ -13,7 +13,6 @@ from transformers_lightning.adapters import SuperAdapter
 from transformers_lightning.datamodules import AdaptersDataModule
 from transformers_lightning.language_modeling.utils import whole_word_tails_mask
 from transformers_lightning.models import TransformersModel
-
 
 transformers.logging.set_verbosity_error()
 logging.getLogger("pytorch_pretrained_bert.tokenization").setLevel(logging.ERROR)