+++ transformers_lightning/loggers/__init__.py
@@ -1,0 +1,1 @@
+from transformers_lightning.loggers.jsonboard_logger import JsonBoardLogger  # noqa: F401