+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 
+## 0.7.10
+
+- Moved `pre_trained_dir` hyperparameter from `Defaults` to `TransformersModelCheckpointCallback`.
+
+- Fixed `JsonboardLogger` with `pytorch-lightning>=1.6`.
+
+
 ## 0.7.9
 
 - Fixed steps computation when `max_steps` is not provided by the user.